:sparkles: upadte Demo.60, update Shape/Circle/Polygon
