+++ src/script/geometry/Circle.ts
@@ -29,10 +29,12 @@ export class Circle extends Shape {
     return polygonCollidesWithCircle(shape as Polygon, this);
   }
 
+  /** @override */
   public getAxes(): Vector[] {
     return undefined;
   }
 
+  /** @override */
   public project(axis: Vector): Projection {
     const scalars = [];
     const point = new Point(this.x, this.y);
@@ -45,13 +47,25 @@ export class Circle extends Shape {
     return new Projection(Math.min.apply(Math, scalars), Math.max.apply(Math, scalars));
   }
 
+  /** @override */
   public move(dx: number, dy: number): void {
     this.x += dx;
     this.y += dy;
   }
 
+  /** @override */
   public createPath(context: CanvasRenderingContext2D): void {
     context.beginPath();
     context.arc(this.x, this.y, this.radius, 0, Math.PI * 2, false);
+  }
+
+  /** @implements */
+  public getClientRect() {
+    return {
+      x: this.x - this.radius,
+      y: this.y - this.radius,
+      width: this.radius * 2,
+      height: this.radius * 2,
+    };
   }
 }
